requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cleanup==8.1.0
django-robots==6.1
django-taggit==6.0.0
lingua-language-detector==2.0.2
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
